Refactor: moved items from .eleventy.js to separate config files
- Moved shortcodes and filters to separate js files in _11ty_config
- Removed browsersync declarations from eleventy config file - it was in there twice for starters, and secondly, the 11ty docs state that it isn't used with the new Eleventy Dev Server which is used in Eleventy 2.0
