+++ .eleventy.js
@@ -2,18 +2,12 @@ const pluginBookshop = require("@bookshop/eleventy-bookshop");
 const yaml = require("js-yaml");
 const svgContents = require("eleventy-plugin-svg-contents");
 const esbuild = require('esbuild');
-const { Tokenizer, assert } = require('liquidjs');
-const path = require("node:path");
-const fs = require('fs'); 
-const Image = require("@11ty/eleventy-img");
 
-const IMAGE_OPTIONS = {
-	widths: [400, 800, 1280, 1600],
-	formats: ["avif", "webp", "svg", "jpeg"],
-	outputDir: "./_site/optimized/",
-	urlPath: "/optimized/",
-	// svgCompressionSize: "br",
-};
+/* 11ty config imports */
+const image_shortcode = require('./_11ty_config/image_shortcode')
+const military_time = require('./_11ty_config/military_time_filter')
+const assign_local_liquid_tag = require('./_11ty_config/assign_local_liquid_tag')
+const contains_block_filter = require('./_11ty_config/contains_block_filter')
 
 const MarkdownIt = require("markdown-it"),
   md = new MarkdownIt({
@@ -32,27 +26,7 @@ module.exports = function (eleventyConfig) {
   eleventyConfig.addDataExtension("yml", (contents) => yaml.load(contents));
 
   // Custom shortcodes
-  eleventyConfig.addShortcode("image", async (srcFilePath, alt, className, sizes, preferSvg) => {
-    let before = Date.now();
-    let inputFilePath = srcFilePath == null ? srcFilePath : path.join(eleventyConfig.dir.input, srcFilePath);
-
-    if (fs.existsSync(inputFilePath)) {
-      let metadata = await Image(inputFilePath, Object.assign({
-        svgShortCircuit: preferSvg ? "size" : false,
-      }, IMAGE_OPTIONS));
-      console.log( `[11ty/eleventy-img] ${Date.now() - before}ms: ${inputFilePath}` );
-
-      return Image.generateHTML(metadata, {
-        alt,
-        class: className,
-        sizes: sizes || "100vw", // Set default value to "100vw" if sizes is not provided
-        loading: "eager",
-        decoding: "async",
-      });
-    } else {
-      return `<img class='${className}' src='${srcFilePath}' alt='${alt}'>`;
-    }
-  });
+  eleventyConfig.addShortcode("image", image_shortcode);
   
   eleventyConfig.addWatchTarget("component-library/");
   
@@ -66,49 +40,11 @@ module.exports = function (eleventyConfig) {
   // Filters
   eleventyConfig.addFilter("markdownify", (markdown) => md.render(markdown));
   eleventyConfig.addFilter("ymlify", (yml) => yaml.load(yml));
-  eleventyConfig.addFilter("militaryTime", function(value) { 
-    const [time, period] = value.split(' '); 
-    const [hour, minute] = time.split(':'); 
-    let formattedHour = parseInt(hour); 
-  
-    if (period === 'pm' || period === 'PM') { 
-        formattedHour += 12; 
-    } 
-  
-    return `${formattedHour}:${minute}`; 
-  });
+  eleventyConfig.addFilter("militaryTime", military_time);
+  eleventyConfig.addFilter('contains_block', contains_block_filter);
 
-  eleventyConfig.setBrowserSyncConfig({
-    files: "./_site/css/**/*.css",
-  });
-
-    // Tags
-    eleventyConfig.addLiquidTag('assign_local', function(liquidEngine) {
-      return {
-        parse: function (token) {
-            const tokenizer = new Tokenizer(token.args, this.liquid.options.operatorsTrie);
-            this.key = tokenizer.readIdentifier().content;
-            tokenizer.skipBlank();
-            assert(tokenizer.peek() === '=', () => `illegal token ${token.getText()}`);
-            tokenizer.advance();
-            this.value = tokenizer.remaining();
-        },
-        render: function(ctx) {
-            ctx.scopes[ctx.scopes.length-1][this.key] = this.liquid.evalValueSync(this.value, ctx);
-        }
-    }
-  });
-
-  eleventyConfig.addFilter('contains_block', function(content_blocks, blockName) {
-    if (!Array.isArray(content_blocks)) {
-      return false;
-    }
-    return content_blocks.some(block => block._bookshop_name === blockName);
-  });
-
-  eleventyConfig.setBrowserSyncConfig({
-    files: './_site/css/**/*.css'
-  });
+  // Tags
+  eleventyConfig.addLiquidTag('assign_local', assign_local_liquid_tag);
 
   // esbuild
   eleventyConfig.addWatchTarget('./src/assets/js/**');
