+++ _11ty_config/military_time_filter.js
@@ -1,0 +1,11 @@
+module.exports = value => { 
+    const [time, period] = value.split(' '); 
+    const [hour, minute] = time.split(':'); 
+    let formattedHour = parseInt(hour); 
+  
+    if (period === 'pm' || period === 'PM') { 
+        formattedHour += 12; 
+    } 
+  
+    return `${formattedHour}:${minute}`
+}
